+++ src/main/php/driver/DummyImageDriver.php
@@ -5,10 +5,10 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  *
- * @package  net\stubbles\img
+ * @package  stubbles\img
  */
-namespace net\stubbles\img\driver;
-use net\stubbles\lang\exception\IOException;
+namespace stubbles\img\driver;
+use stubbles\lang\exception\IOException;
 /**
  * Dummy driver for images.
  */
@@ -83,7 +83,7 @@ class DummyImageDriver implements ImageDriver
      *
      * @return  string
      */
-    public function getLastStoredFileName()
+    public function lastStoredFileName()
     {
         return $this->lastStoredFileName;
     }
@@ -93,7 +93,7 @@ class DummyImageDriver implements ImageDriver
      *
      * @return  resource
      */
-    public function getLastStoredHandle()
+    public function lastStoredHandle()
     {
         return $this->lastStoredHandle;
     }
@@ -123,7 +123,7 @@ class DummyImageDriver implements ImageDriver
      *
      * @return  resource
      */
-    public function getLastDisplayedHandle()
+    public function lastDisplayedHandle()
     {
         return $this->lastDisplayedHandle;
     }
@@ -133,7 +133,7 @@ class DummyImageDriver implements ImageDriver
      *
      * @return  string
      */
-    public function getExtension()
+    public function fileExtension()
     {
         return '.dummy';
     }
@@ -143,7 +143,7 @@ class DummyImageDriver implements ImageDriver
      *
      * @return  string
      */
-    public function getContentType()
+    public function mimeType()
     {
         return 'image/dummy';
     }