+++ src/main/php/response/DefaultImageResponse.php
@@ -5,11 +5,11 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  *
- * @package  net\stubbles\img
+ * @package  stubbles\img
  */
-namespace net\stubbles\img\response;
-use net\stubbles\img\Image;
-use net\stubbles\webapp\response\WebResponse;
+namespace stubbles\img\response;
+use stubbles\img\Image;
+use stubbles\webapp\response\WebResponse;
 /**
  * Response which contains only an image.
  */
@@ -31,7 +31,7 @@ class DefaultImageResponse extends WebResponse implements ImageResponse
     public function setImage(Image $image)
     {
         $this->image = $image;
-        $this->addHeader('Content-type', $image->getContentType());
+        $this->addHeader('Content-type', $image->mimeType());
         return $this;
     }