+++ src/test/php/response/DefaultImageResponseTestCase.php
@@ -5,13 +5,14 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  *
- * @package  net\stubbles\img
+ * @package  stubbles\img
  */
-namespace net\stubbles\img\response;
-use net\stubbles\img\Image;
-use net\stubbles\img\ImageType;
+namespace stubbles\img\response;
+use stubbles\img\Image;
+use stubbles\img\ImageType;
+use stubbles\lang\Rootpath;
 /**
- * Test for net\stubbles\img\response\DefaultImageResponse.
+ * Test for stubbles\img\response\DefaultImageResponse.
  */
 class DefaultImageResponseTestCase extends \PHPUnit_Framework_TestCase
 {
@@ -46,10 +47,11 @@ class DefaultImageResponseTestCase extends \PHPUnit_Framework_TestCase
     public function setUp()
     {
         ImageType::$DUMMY->value()->reset();
-        $this->defaultImageResponse = $this->getMockBuilder('net\stubbles\img\response\DefaultImageResponse')
-                                           ->setMethods(array('header', 'sendBody'))
+        $this->defaultImageResponse = $this->getMockBuilder('stubbles\img\response\DefaultImageResponse')
+                                           ->setMethods(['header', 'sendBody'])
                                            ->getMock();
-        $this->handle               = imagecreatefrompng(\net\stubbles\lang\ResourceLoader::getRootPath() . '/src/test/resources/empty.png');
+        $rootpath                   = new Rootpath();
+        $this->handle               = imagecreatefrompng($rootpath->to('/src/test/resources/empty.png'));
         $this->image                = new Image('test', ImageType::$DUMMY, $this->handle);
     }
 
@@ -70,7 +72,7 @@ class DefaultImageResponseTestCase extends \PHPUnit_Framework_TestCase
     {
         $this->defaultImageResponse->expects($this->at(1))
                                    ->method('header')
-                                   ->with($this->equalTo('Content-type: ' . ImageType::$DUMMY->getContentType()));
+                                   ->with($this->equalTo('Content-type: ' . ImageType::$DUMMY->mimeType()));
         $this->defaultImageResponse->setImage($this->image)
                                    ->send();
     }
@@ -82,7 +84,7 @@ class DefaultImageResponseTestCase extends \PHPUnit_Framework_TestCase
     {
         $this->defaultImageResponse->setImage($this->image)->send();
         $this->assertSame($this->handle,
-                          ImageType::$DUMMY->value()->getLastDisplayedHandle()
+                          ImageType::$DUMMY->value()->lastDisplayedHandle()
         );
     }
 
@@ -93,7 +95,7 @@ class DefaultImageResponseTestCase extends \PHPUnit_Framework_TestCase
     public function clearRemovesImageFromResponse()
     {
         $this->defaultImageResponse->setImage($this->image)->clear()->send();
-        $this->assertNull(ImageType::$DUMMY->value()->getLastDisplayedHandle());
+        $this->assertNull(ImageType::$DUMMY->value()->lastDisplayedHandle());
     }
 
     /**