+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+3.0.0 (2014-06-??)
+------------------
+
+### BC breaks
+
+  * removed namespace prefix `net`, base namespace is now `stubbles\img` only
+  * api changes
+    * renamed methods in `stubbles\image\driver\ImageDriver`:
+      * `getExtension()` is now `fileExtension()`
+      * `getContentType()` is now `mimeType()`
+    * renamed methods in `stubbles\image\driver\DummyImageDriver`:
+      * `getLastStoredFileName()` is now `lastStoredFileName()`
+      * `getLastStoredHandle()` is now `lastStoredHandle()`
+      * `getLastDisplayedHandle()` is now `lastDisplayedHandle()`
+    * deprecated `stubbles\img\Image::getExtension()`, use `stubbles\img\Image::fileExtension()` instead, will be removed with 4.0.0
+    * deprecated `stubbles\img\Image::getContentType()`, use `stubbles\img\Image::mimeType()` instead, will be removed with 4.0.0
+
+### Other changes
+
+  * upgraded to stubbles/core 4.x & stubbles/webapp-core 4.x
+
+
 2.0.3 (2013-11-04)
 ------------------